tic-tac-toe/demo/app.py

import gradio as gr
from gradio_tic-tac-toe import tic-tac-toe


with gr.Blocks() as demo:
    gr.Markdown("# Change the value (keep it JSON) and the front-end will update automatically.")
    tic-tac-toe(value={"message": "Hello from Gradio!"}, label="Static")


if __name__ == "__main__":
    demo.launch()
